pagination added

+++ src/Formularies.UserManagementService.Api/V1/Controllers/UsersController.cs
@@ -1,4 +1,7 @@
-﻿using Formularies.UserManagementService.Core.Interfaces.Services;
+﻿using AutoWrapper.Extensions;
+using AutoWrapper.Wrappers;
+using Formularies.UserManagementService.Core.Helpers;
+using Formularies.UserManagementService.Core.Interfaces.Services;
 using Formularies.UserManagementService.Core.Models;
 using Formularies.UserManagementService.Core.Request;
 using Formularies.UserManagementService.Core.Response;
@@ -6,21 +9,25 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Formularies.UserManagementService.Api.V1.Controllers
 {
     [Produces("application/json")]
-    [Route("api/" + ApiConstants.ServiceName + "/v{api-version:apiVersion}/[controller]")]
+    [Route("api/v{api-version:apiVersion}/[controller]")]
     [ApiVersion("1.0")]
     [ApiController]
     public class UsersController : ControllerBase
     {
         private readonly IUserService _userService;
+        private readonly IUriService _uriService;
+        public new User User => (User)HttpContext.Items["User"];
 
-        public UsersController(IUserService userService)
+        public UsersController(IUserService userService, IUriService uriService)
         {
             _userService = userService ?? throw new ArgumentNullException(nameof(userService));
+            _uriService = uriService ?? throw new ArgumentNullException(nameof(uriService));
         }
 
         /// <summary>
@@ -30,16 +37,37 @@ namespace Formularies.UserManagementService.Api.V1.Controllers
         /// <remarks>
         /// Tables used => Users
         /// </remarks>
+        //[HttpGet(Name = "GetUsers")]
+        //[ProducesResponseType(StatusCodes.Status200OK)]
+        //[ProducesResponseType(StatusCodes.Status404NotFound)]
+        //[ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        //[ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        //public async Task<ActionResult<IEnumerable<User>>> GetUsers()
+        //{
+        //    var response = await _userService.GetAllUsers().ConfigureAwait(false);
+        //    return response != null ? Ok(new Response<IEnumerable<User>>(response)) : NotFound();
+        //}
+
         [HttpGet(Name = "GetUsers")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<IEnumerable<User>>> GetUsers()
+        public ActionResult<IEnumerable<User>> GetUsers([FromQuery] PaginationFilter pageFilter, [FromQuery] SearchFilter searchFilter)
         {
-            var response = await _userService.GetAllUsers().ConfigureAwait(false);
-            return response != null ? Ok(response) : NotFound();
+            var route = Request.Path.Value;
+            var response = _userService.GetAllUsers(searchFilter);
+            var validFilter = new PaginationFilter(pageFilter.PageNumber, pageFilter.PageSize);
+            var pagedData = response
+               .Skip((validFilter.PageNumber - 1) * validFilter.PageSize)
+               .Take(validFilter.PageSize)
+               .ToList();
+            var totalRecords = response.Count();
+            var pagedReponse = PaginationHelper.CreatePagedReponse<User>(pagedData, validFilter, totalRecords, _uriService, route);
+            return pagedReponse != null ? Ok(pagedReponse) : NotFound();
         }
+
+
 
         /// <summary>
         /// Get user by id
@@ -56,9 +84,13 @@ namespace Formularies.UserManagementService.Api.V1.Controllers
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<User>> GetUserById(Guid id)
-        {            
+        {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             var response = await _userService.GetUserById(id).ConfigureAwait(false);
-            return response != null ? Ok(response) : NotFound();
+            return response != null ? Ok(new ApiResponse("Get request successful", response, 200)) : NotFound();
 
         }
 
@@ -79,14 +111,18 @@ namespace Formularies.UserManagementService.Api.V1.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<User>> CreateUser(User User)
+        public async Task<ActionResult<UserCreateRequest>> CreateUser(UserCreateRequest User)
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest();
+                throw new ApiException(ModelState.AllErrors());
+            }
+            if (User.RoleId <= 0)
+            {
+                throw new ApiException("Role id not valid", 400);
             }
             var response = await _userService.CreateUser(User).ConfigureAwait(false);
-            return CreatedAtRoute(nameof(GetUserById), new { id = response.UserId }, response);
+            return CreatedAtRoute(nameof(GetUserById), new { id = response.UserId }, new ApiResponse("Post request successful", response, 201));
         }
 
         /// <summary>
@@ -104,7 +140,11 @@ namespace Formularies.UserManagementService.Api.V1.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<bool>> DeleteUser(Guid id)
-        {           
+        {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             var response = await _userService.DeleteUser(id).ConfigureAwait(false);
             return response ? NoContent() : NotFound();
         }
@@ -125,69 +165,81 @@ namespace Formularies.UserManagementService.Api.V1.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<bool>> UpdateUser(Guid id, User User)
+        public async Task<ActionResult<bool>> UpdateUser(Guid id, UserUpdateRequest User)
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest();
+                throw new ApiException(ModelState.AllErrors());
+            }
+            if (User.RoleId <= 0)
+            {
+                throw new ApiException("Role id not valid", 400);
             }
             var response = await _userService.UpdateUser(id, User).ConfigureAwait(false);
-            return response ? Ok(response) : NotFound();
+            return response ? Ok(new ApiResponse("Put request successful", response, 200)) : NotFound();
         }
 
         [HttpPost("Authenticate")]
-        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<AuthenticateResponse>> Authenticate(AuthenticateRequest model)
         {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             var response = await _userService.Authenticate(model, ipAddress());
             setTokenCookie(response.RefreshToken);
-            return Ok(response);
+            return Ok(new ApiResponse("Post request successful", response, 200));
         }
 
         [HttpPost("RefreshToken")]
-        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<AuthenticateResponse>> RefreshToken()
         {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             var refreshToken = Request.Cookies["refreshToken"];
             var response = await _userService.RefreshToken(refreshToken, ipAddress());
             setTokenCookie(response.RefreshToken);
-            return Ok(response);
+            return Ok(new ApiResponse("Post request successful", response, 200));
         }
 
         [HttpPost("ForgotPassword")]
-        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> ForgotPassword(ForgotPasswordRequest model)
         {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             await _userService.ForgotPassword(model, Request.Headers["origin"]);
-            return Ok(new { message = "Please check your email for password reset instructions" });
+            return Ok(new ApiResponse("Forgot passsword successful, please check your email for password reset instructions", 200));
         }
 
         [HttpPost("ResetPassword")]
-        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> ResetPassword(ResetPasswordRequest model)
         {
+            if (!ModelState.IsValid)
+            {
+                throw new ApiException(ModelState.AllErrors());
+            }
             await _userService.ResetPassword(model);
-            return Ok(new { message = "Password reset successful, you can now login" });
+            return Ok(new ApiResponse("Password reset successful, you can now login",200));
         }
 
         private void setTokenCookie(string token)
